+++ Section.h
@@ -22,9 +22,9 @@ public:
 		"TRUE"
 	};
 private:
-	unsigned m_StartAddress;	//where section starts in memory
-	unsigned m_Size;		//max size of the memory block
-	unsigned m_LocationCounter;	//current location to store data
+	int m_StartAddress;	//where section starts in memory
+	int m_Size;		//max size of the memory block
+	int m_LocationCounter;	//current location to store data
 	char* m_pSectionData;	//where the data is
 	Mode m_AccessMode;	//read only or read write
 	AddressSize m_ZeroPageAddressSize;
@@ -32,21 +32,21 @@ public:
 	CSection();
 	virtual ~CSection(); 
 	bool Create();
-	unsigned GetStartAddress() { return m_StartAddress; }
-	void SetStartAddress(unsigned SA);
-	unsigned GetSectionSize() { return m_Size; }
-	void SetSectionSize(unsigned S);
-	unsigned short GetLocationCounter() { return m_LocationCounter; }
-	unsigned short IncrementLocationCounterBy(unsigned short v) { 
+	int GetStartAddress() { return m_StartAddress; }
+	void SetStartAddress(int SA);
+	int GetSectionSize() { return m_Size; }
+	void SetSectionSize(int S);
+	int GetLocationCounter() { return m_LocationCounter; }
+	int IncrementLocationCounterBy(int v) {
 		m_LocationCounter += v;
 		return m_LocationCounter;
 	}
-	void SetLocationCounter(unsigned short NewAddress);
-	unsigned short AddInstruction(CAct65Opcode* pINS);
-	unsigned short AddData(unsigned ObjectSize, int Value);
-	unsigned short AddData(unsigned ObjSize, const char* pData);
+	void SetLocationCounter(int NewAddress);
+	int AddInstruction(CAct65Opcode* pINS);
+	int AddData(int ObjectSize, int Value);
+	int AddData(int ObjSize, const char* pData);
 	void AddDataAt(unsigned Adr, unsigned ObjectSize, int valu);
-	unsigned short AllocateDataBlock(unsigned short size);	//define Storage
+	int AllocateDataBlock(int size);	//define Storage
 	void SetAccessMode(Mode Am) { m_AccessMode = Am; }
 	CSection::Mode GetAccessMode() { return m_AccessMode; }
 	void SetZeroPageFlag(AddressSize SizeFlag) { m_ZeroPageAddressSize = SizeFlag; }
